feat: Add adaptive pause functionality to block job
This commit adds adaptive pause functionality to the block job. It introduces a new function, block_job_adaptive_pause(), which takes in the job, the BlockStateDriver, the threshold, and the pause time as parameters, to implement adaptive stream pause based on the provided (or calculated) threshold.

The IO tracking functions have been moved to block-backend.c, and byte read/write counter is moved to blk_aio_write/read_entry(), which is the entry point for IO done by the guest.

In stream.c, the IO tracker is initialised in cor_filter_bs, which seems to hold the block backend where the IO by the guest is being handled. This essentialy rules out counting IO which is not triggered by the guest.